Introduce DiceTestTrait for partial mocking DI:: calls
